[MSLUGSBF] Make 64bit

+++ ports/released/gamemakerengine/metalslugsbfanthology/Metal Slug SB Fanthology.sh
@@ -13,58 +13,48 @@ else
 fi
 
 source $controlfolder/control.txt
-source $controlfolder/device_info.txt
 [ -f "${controlfolder}/mod_${CFW_NAME}.txt" ] && source "${controlfolder}/mod_${CFW_NAME}.txt"
 get_controls
-
-# Echo
-echo "Loading port, please wait..." > $CUR_TTY
-
-# Setup permissions
-$ESUDO chmod 666 /dev/tty1
-$ESUDO chmod 666 /dev/uinput
 
 # Variables
 GAMEDIR="/$directory/ports/mslugsbf"
 
+# CD and set logging
 cd $GAMEDIR
 > "$GAMEDIR/log.txt" && exec > >(tee "$GAMEDIR/log.txt") 2>&1
 
-$ESUDO chmod 777 "$GAMEDIR/gmloadernext-armhf"
+# Setup permissions
+$ESUDO chmod +xwr "$GAMEDIR/gmloadernext.aarch64"
+$ESUDO chmod +xr "$GAMEDIR/tools/splash"
 
 # Exports
-export LD_LIBRARY_PATH="/usr/lib:/usr/lib32:/$GAMEDIR/libs:$LD_LIBRARY_PATH"
-export GMLOADER_PLATFORM="os_windows"
-export PORT_32BIT="Y"
 export SDL_GAMECONTROLLERCONFIG="$sdl_controllerconfig"
 
-install-win() {
-    if [ ! -f "$GAMEDIR/installed" ]; then	
-        echo "Performing first-run setup..." > $CUR_TTY
-        # Purge unneeded files
-        rm -rf assets/*.ini assets/*.exe assets/*.dll assets/*.png || exit 1
-        # Rename data.win
-        echo "Moving the game file..." > $CUR_TTY
-        mv "./assets/data.win" "./game.droid" || exit 1
-        # Create a new zip file game.apk from specified directories
-        echo "Zipping assets into apk..." > $CUR_TTY
-        ./utils/zip -r -0 "game.apk" "assets" || exit 1
-        # Remove assets directory
-        rm -rf "$GAMEDIR/assets" || exit 1
-        # Mark installation as complete
-        touch "$GAMEDIR/installed"
+# Check if we need to patch the game
+if [ ! -f patchlog.txt ] || [ -f "$GAMEDIR/assets/data.win" ]; then
+    if [ -f "$controlfolder/utils/patcher.txt" ]; then
+        export PATCHER_FILE="$GAMEDIR/tools/patchscript"
+        export PATCHER_GAME="$(basename "${0%.*}")"
+        export PATCHER_TIME="2 to 5 minutes"
+        export controlfolder
+        export ESUDO
+        source "$controlfolder/utils/patcher.txt"
+        $ESUDO kill -9 $(pidof gptokeyb)
+    else
+        echo "This port requires the latest version of PortMaster."
     fi
-}
+fi
 
-if [ ! -f "$GAMEDIR/installed" ]; then
-    install-win
+# Display loading splash
+if [ -f "$GAMEDIR/patchlog.txt" ]; then
+    [ "$CFW_NAME" == "muOS" ] && $ESUDO "$GAMEDIR/tools/splash" "$GAMEDIR/splash.png" 1
+    $ESUDO "$GAMEDIR/tools/splash" "$GAMEDIR/splash.png" 8000 & 
 fi
 
 # Assign gptokeyb and load the game
-$GPTOKEYB "gmloadernext-armhf" -c "control.gptk" &
-./gmloadernext-armhf game.apk
+$GPTOKEYB "gmloadernext.aarch64" -c "mslug.gptk" &
+pm_platform_helper "$GAMEDIR/gmloadernext.aarch64" >/dev/null
+./gmloadernext.aarch64 -c gmloader.json
 
-# Kill processes
-$ESUDO kill -9 $(pidof gptokeyb)
-$ESUDO systemctl restart oga_events &
-printf "\033c" > /dev/tty0
+# Cleanup
+pm_finish
